requirements.txt
llama-index==0.10.23
llama-hub==0.0.79
python-dotenv==1.0.1
tiktoken==0.6.0
streamlit==1.32.2
